fix jackson 2.5.x support
jackson 2.5.x calls `AnnotationIntrospector#hasCreatorAnnotation()`
to verify that a constructor should be used in deserialization. This
does not seem to have been done by 2.4.x. As such it looks like a
breaking bug fix change in jackson. The default implementation in
NopAnnotationIntrospector as extended by
AutoMatterAnnotationIntrospector returns false.

To fix this, we implement `#hasCreatorAnnotation()` and have it return
`true` if there's either no fields in the value or the first field is
annotated with `@AutoMatter.Field`.

Also, bump the jackson dependency to 2.5.3.