Use the new type adapter
The type adapter factory is now smarter and picks a specialized
type adapter if it finds @SerializedName annotations on the interface
or on any of the interfaces implemented by the value class, if not
it delegates to gson to pick a factory to use. This means a slighlty
slower path for any annotated classes (time spent replacing fields
increase linearly with the amount of annotations used).