Eagerly calculate the translated fieldname
This should increase performance as the name is now only translated
once instead of everytime we go through read or write path.

As the method is moved to the TypeAdapterFactory that is kept in
memory for the life time of the user created gson instance, the gson
instance passed to the create method is forwarded to the methods that
calculate the annotations and translates the fieldname. This decreases
the amount of memory and increases speed of access to the gson object
passed to us.