Ignore default methods that are also present on superinterfaces
Currently if you have a default method on an AutoMatter annotated
interface Foo that is also present in Bar (and Foo extends Bar)
AutoMatter will treat the method as a field. The motivation for that
setup is for when you want the generated values to implement some
interface.

This patch solves the issue by enumerating methods into a map so that
any default methods defined in the interface will take precedence over
what's present in the superinterfaces.

A `LinkedHashMap` is used to preserve insertion order.