Don't emit `@SafeVarargs` for reifiable types
Attempt to detect if a type is reifiable, and omit the annotation in
those cases.

https://docs.oracle.com/javase/specs/jls/se7/html/jls-4.html

However one problem as noted in the comments is we can't get at a
ParameterizedTypeName's enclosing type. So for instance, this code will
treat `Foo<X>.Bar` as reifiable, even though it isn't.

One idea to get around it would be to `toString()` the type and check if
it contains any type parameters, although that seems kinda messy.